+++ apps/meteor/app/livechat/server/lib/guests.ts
@@ -1,5 +1,5 @@
 import { Apps, AppEvents } from '@rocket.chat/apps';
-import type { ILivechatVisitor } from '@rocket.chat/core-typings';
+import type { ILivechatVisitor, IOmnichannelRoom } from '@rocket.chat/core-typings';
 import {
 	LivechatVisitors,
 	LivechatCustomField,
@@ -9,13 +9,16 @@ import {
 	ReadReceipts,
 	Subscriptions,
 	LivechatContacts,
+	Users,
 } from '@rocket.chat/models';
 import { wrapExceptions } from '@rocket.chat/tools';
+import UAParser from 'ua-parser-js';
 
-import { validateEmail } from './Helper';
+import { parseAgentCustomFields, validateEmail } from './Helper';
 import type { RegisterGuestType } from './Visitors';
 import { Visitors } from './Visitors';
 import { ContactMerger, type FieldAndValue } from './contacts/ContactMerger';
+import type { ICRMData } from './localTypes';
 import { livechatLogger } from './logger';
 import { trim } from '../../../../lib/utils/stringUtils';
 import { i18n } from '../../../../server/lib/i18n';
@@ -159,4 +162,65 @@ async function cleanGuestHistory(token: string) {
 	const livechatInquiries = await LivechatInquiry.findIdsByVisitorToken(token).toArray();
 	await LivechatInquiry.removeByIds(livechatInquiries.map(({ _id }) => _id));
 	void notifyOnLivechatInquiryChanged(livechatInquiries, 'removed');
+}
+
+export async function getLivechatRoomGuestInfo(room: IOmnichannelRoom) {
+	const visitor = await LivechatVisitors.findOneEnabledById(room.v._id);
+	if (!visitor) {
+		throw new Error('error-invalid-visitor');
+	}
+
+	const agent = room.servedBy?._id ? await Users.findOneById(room.servedBy?._id) : null;
+
+	const ua = new UAParser();
+	ua.setUA(visitor.userAgent || '');
+
+	const postData: ICRMData = {
+		_id: room._id,
+		label: room.fname || room.label, // using same field for compatibility
+		topic: room.topic,
+		createdAt: room.ts,
+		lastMessageAt: room.lm,
+		tags: room.tags,
+		customFields: room.livechatData,
+		visitor: {
+			_id: visitor._id,
+			token: visitor.token,
+			name: visitor.name,
+			username: visitor.username,
+			department: visitor.department,
+			ip: visitor.ip,
+			os: ua.getOS().name && `${ua.getOS().name} ${ua.getOS().version}`,
+			browser: ua.getBrowser().name && `${ua.getBrowser().name} ${ua.getBrowser().version}`,
+			customFields: visitor.livechatData,
+		},
+	};
+
+	if (agent) {
+		const customFields = parseAgentCustomFields(agent.customFields);
+
+		postData.agent = {
+			_id: agent._id,
+			username: agent.username,
+			name: agent.name,
+			...(customFields && { customFields }),
+		};
+
+		if (agent.emails && agent.emails.length > 0) {
+			postData.agent.email = agent.emails[0].address;
+		}
+	}
+
+	if (room.crmData) {
+		postData.crmData = room.crmData;
+	}
+
+	if (visitor.visitorEmails && visitor.visitorEmails.length > 0) {
+		postData.visitor.email = visitor.visitorEmails;
+	}
+	if (visitor.phone && visitor.phone.length > 0) {
+		postData.visitor.phone = visitor.phone;
+	}
+
+	return postData;
 }
