+++ apps/meteor/app/livechat/server/lib/localTypes.ts
@@ -1,4 +1,12 @@
-import type { IOmnichannelRoom, IUser, ILivechatVisitor, IMessage, MessageAttachment, IMessageInbox } from '@rocket.chat/core-typings';
+import type {
+	IOmnichannelRoom,
+	IUser,
+	ILivechatVisitor,
+	IMessage,
+	MessageAttachment,
+	IMessageInbox,
+	IOmnichannelAgent,
+} from '@rocket.chat/core-typings';
 
 type GenericCloseRoomParams = {
 	room: IOmnichannelRoom;
@@ -54,3 +62,23 @@ export interface ILivechatMessage {
 	blocks?: IMessage['blocks'];
 	email?: IMessageInbox['email'];
 }
+
+export type ICRMData = {
+	_id: string;
+	label?: string;
+	topic?: string;
+	createdAt: Date;
+	lastMessageAt?: Date;
+	tags?: string[];
+	customFields?: IOmnichannelRoom['livechatData'];
+	visitor: Pick<ILivechatVisitor, '_id' | 'token' | 'name' | 'username' | 'department' | 'phone' | 'ip'> & {
+		email?: ILivechatVisitor['visitorEmails'];
+		os?: string;
+		browser?: string;
+		customFields: ILivechatVisitor['livechatData'];
+	};
+	agent?: Pick<IOmnichannelAgent, '_id' | 'username' | 'name' | 'customFields'> & {
+		email?: NonNullable<IOmnichannelAgent['emails']>[number]['address'];
+	};
+	crmData?: IOmnichannelRoom['crmData'];
+};
